Split plots.ts handlers + tests into per-domain files (warren-332b)
Phase 2 of the plots handler split: delete the 1010-line
src/server/handlers/plots.ts and the 2399-line plots.test.ts, moving
the remaining handlers into plots/ domain files:

- plots/shared.ts   — resolvePlotProject / buildPlotEnvelope / loadPausedRunsForPlot
- plots/detail.ts   — GET /plots/:id, GET /plots/:id/summary
- plots/intent.ts   — POST /plots/:id/intent, POST /plots/:id/rename
- plots/status.ts   — POST /plots/:id/status
- plots/attachments.ts — attach / detach / merge-PR

plots/index.ts is now a pure re-export composer; sync.ts reuses
resolvePlotProject. The monolithic suite is split into
plots.{detail,intent,rename,status,attach,detach,merge}.test.ts, all
under the 500-line budget, sharing plots.test-support.ts.

attach/merge handlers gained parseAttachmentBody / parseMergeMethod
helpers to fall under cognitive-complexity 15, so the stale
src/server/handlers/plots.ts biome grandfather entry is removed (not
relocated), along with the plots.ts + plots.test.ts file-size budgets.

Wire contract is byte-identical. bun run check:all is green.
