+++ .gitconfig.burrow
@@ -1,0 +1,3 @@
+[user]
+	name = Warren
+	email = 28354219+jayminwest@users.noreply.github.com
