+++ .mulch/expertise/build.jsonl
@@ -94,3 +94,4 @@
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-29T02:13:34.165Z","evidence":{"seeds":"warren-6376","commit":"2a0483bc662b04aae9a485189af4cbec5ab7ea2a"},"content":"src/server/bridge-reconnect.ts (extracted from bridges.ts to stay under the file-size ratchet) holds runWithReconnect: after BRIDGE_STALL_THRESHOLD=3 consecutive errored reconnects with no forward progress it emits a one-shot bridge_stalled (stream:system) event, and bridge_recovered once events stream again (result.written>0 resets the counter + clears the stall). UI: RunDetail.tsx isBridgeStalled() scans the events tail and renders a warning Alert banner while !isTerminal. Shared bridge test fixtures now live in src/server/bridges.test-helpers.ts. Gotcha: the raw-JS bundle budget (584000) sits at its temporary pl-55a3 ceiling with ~0 headroom, so the banner only fit after trimming dev tracker/SPEC refs from user-facing title= tooltips.","id":"mx-9eb278"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-29T02:25:02.911Z","evidence":{"seeds":"warren-bfc6","commit":"3de3b60267e54ab386f91de650cb125371032d22"},"relates_to":["mx-f1f5f7"],"content":"Bundle-size ratchet close-out (warren-bfc6 / pl-369d): after pl-55a3 (warren-6358) closed, the temporary overhaul ceilings in scripts/bundle-size-budgets.json were collapsed back to the real frozen floor. Re-measured from clean 'bun run build:ui' + 'bun run check:bundle-size': raw js 583998 / raw css 40432 / gzip js 172380 / gzip css 7942 (single js+css asset so largest==totals). Budgets set to measured + ~150-200B headroom; stacked comment_phase notes collapsed into one closeout note. Bulk of the js floor is the Phase 5 framer-motion v12 (strict LazyMotion + domAnimation) layer.","id":"mx-2b7918"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-29T04:19:53.762Z","evidence":{"seeds":"warren-e577","commit":"54dd10209846d01516ca17cdf31b6bbccf1f25c8"},"content":"warren-e577 / pl-369d: updated ~13 stale 'src/server/main.ts' references to 'src/server/main/index.ts' across SPEC.md, src/ui/README.md and 7 comment headers (handlers/index, plots/aggregate, plots/errors, preview/launch/url, db/client, cli/commands/serve, runs/pause) after v0.7.0 PR #181 directory split. Doc/comment-only; check:all coverage gate red on pre-existing functions floor tracked by warren-88da, unaffected by this change.","id":"mx-4112d9"}
+{"type":"convention","classification":"tactical","recorded_at":"2026-05-29T04:39:57.179Z","evidence":{"seeds":"warren-332b","commit":"d0f853991a8112e78331cae6d1c7ff412257f6ab"},"relates_to":["mx-3df5c5"],"content":"Plots handlers split phase 2 (warren-332b / pl-369d): src/server/handlers/plots.ts (1010 lines) deleted; remaining handlers extracted into plots/ domain files — detail.ts (GET /plots/:id + /summary), intent.ts (intent + rename), status.ts (status), attachments.ts (attach/detach/merge-PR). Shared helpers in plots/shared.ts: resolvePlotProject(deps,plotId,action) throws byte-identical 404/ProjectLacksPlotError; buildPlotEnvelope; loadPausedRunsForPlot. plots/index.ts is now a pure re-export composer. plots.test.ts (2399 lines) split into plots.{detail,intent,rename,status,attach,detach,merge}.test.ts sharing plots.test-support.ts (non-.test.ts helper). attach/merge handlers gained parseAttachmentBody/parseMergeMethod helpers to drop under cognitive-complexity 15, so the old plots.ts biome grandfather entry was removed (not relocated); file-size budgets for plots.ts + plots.test.ts removed.","id":"mx-ed7861"}
